ssh_honeypot/requirements.txt
paramiko>=3.4.0
pymongo>=4.6.0
motor>=3.3.2
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99
Pillow>=10.2.0
cryptography>=41.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
aiohttp>=3.9.0
openai>=1.10.0
textblob>=0.18.0
vaderSentiment>=3.3.2
numpy>=1.26.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
